requirements.txt
mcp>=0.9.0
pymongo>=4.6.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0

